pyproject.toml
[project]
name = "opensearch-mcp-server"
version = "1.0.0"
description = "MCP Server for interacting with OpenSearch"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opensearch-py<=2.8.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.4.0",
]

[project.license]
file = "LICENSE"

[project.scripts]
opensearch-mcp-server = "opensearch_mcp_server:main"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
